E2E: LoginInteraction moved to interactions directory (resolves #805)
