+++ ee_tests/src/booster_ee_int_tests/booster_build_maven.spec.ts
@@ -2,6 +2,7 @@ import { browser, ExpectedConditions as until } from 'protractor';
 
 import * as support from '../support';
 import { Button } from '../ui/button';
+import { LoginInteraction } from '../interactions/login_interactions';
 
 import { MainDashboardPage } from '../page_objects/main_dashboard.page';
 import { SpaceCheWorkspacePage } from '../page_objects/space_cheworkspace.page';
@@ -11,7 +12,7 @@ describe('Build the project using Maven in the Che terminal window', () => {
 
   beforeEach(async () => {
     await support.desktopTestSetup();
-    let login = new support.LoginInteraction();
+    let login = new LoginInteraction();
     await login.run();
     dashboardPage = new MainDashboardPage();
   });
