+++ ee_tests/src/booster_ee_int_tests/booster_oso_project.spec.ts
@@ -2,6 +2,7 @@ import { browser, ExpectedConditions as until } from 'protractor';
 
 import * as support from '../support';
 import { Button } from '../ui/button';
+import { LoginInteraction } from '../interactions/login_interactions';
 
 import { SpaceDashboardPage } from '../page_objects/space_dashboard.page';
 import { SpacePipelinePage } from '../page_objects/space_pipeline_tab.page';
@@ -14,7 +15,7 @@ describe('Access project in OSO:', () => {
 
   beforeEach(async () => {
     await support.desktopTestSetup();
-    let login = new support.LoginInteraction();
+    let login = new LoginInteraction();
     await login.run();
     dashboardPage = new MainDashboardPage();
   });
