+++ ee_tests/src/booster_ee_int_tests/booster_setup.spec.ts
@@ -9,6 +9,7 @@ import { SpaceChePage } from '../page_objects/space_che.page';
 
 import { SpaceDashboardInteractionsFactory } from '../interactions/space_dashboard_interactions';
 import { AccountHomeInteractionsFactory } from '../interactions/account_home_interactions';
+import { LoginInteraction } from '../interactions/login_interactions';
 import { PageOpenMode } from '../page_objects/base.page';
 import { Button } from '../ui/button';
 
@@ -19,7 +20,7 @@ describe('Creating new quickstart in OSIO', () => {
 
   beforeEach(async () => {
     await support.desktopTestSetup();
-    let login = new support.LoginInteraction();
+    let login = new LoginInteraction();
     await login.run();
     dashboardPage = new MainDashboardPage();
   });
