+++ ee_tests/src/booster_ee_int_tests/booster_trigger_cd.spec.ts
@@ -2,6 +2,7 @@ import {
   browser, by, element, ElementFinder, ExpectedConditions as until,
   Key, protractor
 } from 'protractor';
+import { LoginInteraction } from '../interactions/login_interactions';
 import * as support from '../support';
 import { Quickstart } from '../support/quickstart';
 import { Button } from '../ui/button';
@@ -32,7 +33,7 @@ describe('Triggers the CD Build (Jenkins):', () => {
 
   beforeEach(async () => {
     await support.desktopTestSetup();
-    let login = new support.LoginInteraction();
+    let login = new LoginInteraction();
     await login.run();
     dashboardPage = new MainDashboardPage();
   });
