+++ ee_tests/src/booster_ee_int_tests/quickstart_che_editor.spec.ts
@@ -1,4 +1,5 @@
 import { browser, ExpectedConditions as until, Key, protractor } from 'protractor';
+import { LoginInteraction } from '../interactions/login_interactions';
 import * as support from '../support';
 import { Quickstart } from '../support/quickstart';
 import { Button } from '../ui/button';
@@ -81,7 +82,7 @@ describe('Creating new Che workspace and edit in OSIO', () => {
 
   beforeEach(async () => {
     await support.desktopTestSetup();
-    let login = new support.LoginInteraction();
+    let login = new LoginInteraction();
     await login.run();
     dashboardPage = new MainDashboardPage();
