+++ ee_tests/src/booster_pipeline.spec.ts
@@ -2,6 +2,7 @@ import { browser, by, element, ElementFinder } from 'protractor';
 
 import * as support from './support';
 import { Button } from './ui/button';
+import { LoginInteraction } from './interactions/login_interactions';
 
 import { SpacePipelinePage } from './page_objects/space_pipeline_tab.page';
 import { MainDashboardPage } from './page_objects/main_dashboard.page';
@@ -40,7 +41,7 @@ describe('Creating new quickstart in OSIO', () => {
 
   beforeEach(async () => {
     await support.desktopTestSetup();
-    let login = new support.LoginInteraction();
+    let login = new LoginInteraction();
     await login.run();
     dashboardPage = new MainDashboardPage();
   });
