+++ ee_tests/src/smoke.spec.ts
@@ -4,6 +4,7 @@ import * as support from './support';
 import { FeatureLevelUtils } from './support/feature_level';
 import { Quickstart } from './support/quickstart';
 import { DeploymentsInteractions, DeploymentsInteractionsFactory } from './interactions/deployments_interactions';
+import { LoginInteraction } from './interactions/login_interactions';
 import { PipelinesInteractions } from './interactions/pipelines_interactions';
 import { SpaceDashboardInteractionsFactory } from './interactions/space_dashboard_interactions';
 import { AccountHomeInteractionsFactory } from './interactions/account_home_interactions';
@@ -51,7 +52,7 @@ describe('e2e_smoketest', () => {
 
   it('login', async () => {
     support.info('--- Login ---');
-    let login = new support.LoginInteraction();
+    let login = new LoginInteraction();
     await login.run();
   });
