+++ ee_tests/src/support/index.ts
@@ -1,2 +1,1 @@
 export * from './core';
-export * from './login_interactions';
